refactor: address review feedback on PR #3350
- Move retrieveDescriptor(eservice, descriptorId) helper into
  notification-commons/retrievers.ts; drop the duplicated local copy in
  both in-app archiving handlers.
- Replace the local "formatItalianDate" helpers (both in inAppTemplates
  and in the email dispatcher's archivingDateFormat.ts) with the
  existing dateAtRomeZone from pagopa-interop-commons, which formats
  dd/MM/yyyy in Europe/Rome timezone. Delete the now-unused
  archivingDateFormat.ts file.
- Drop redundant inline comments in templates.ts, inAppTemplates.ts and
  handleEserviceArchivingToConsumer.ts (the code is self-explanatory).
