--- packages/email-notification-dispatcher/src/handlers/eservices/archivingDateFormat.ts
@@ -1,6 +1,0 @@
-export const formatItalianDate = (date: Date): string =>
-  new Intl.DateTimeFormat("it-IT", {
-    day: "2-digit",
-    month: "2-digit",
-    year: "numeric",
-  }).format(date);
