+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceArchivingScheduledToConsumer.ts
@@ -16,7 +16,7 @@ import {
 } from "pagopa-interop-notification-commons";
 import { EServiceHandlerParams } from "../handlerCommons.js";
 import { config } from "../../config/config.js";
-import { formatItalianDate } from "./archivingDateFormat.js";
+import { dateAtRomeZone } from "pagopa-interop-commons";
 
 const notificationType: NotificationType = "eserviceStateChangedToConsumer";
 
@@ -70,7 +70,7 @@ export async function handleEserviceArchivingScheduledToConsumer(
   }
 
   const archivableOn = descriptor.archivingSchedule
-    ? formatItalianDate(descriptor.archivingSchedule.archivableOn)
+    ? dateAtRomeZone(descriptor.archivingSchedule.archivableOn)
     : undefined;
   const subject = `Avvio archiviazione dell'e-service "${eservice.name}"`;
