+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceDescriptorArchivedToProducer.ts
@@ -17,7 +17,7 @@ import {
 } from "pagopa-interop-notification-commons";
 import { EServiceDescriptorHandlerParams } from "../handlerCommons.js";
 import { config } from "../../config/config.js";
-import { formatItalianDate } from "./archivingDateFormat.js";
+import { dateAtRomeZone } from "pagopa-interop-commons";
 
 const notificationType: NotificationType = "eserviceStateChangedToProducer";
 
@@ -78,7 +78,7 @@ export async function handleEserviceDescriptorArchivedToProducer(
   }
 
   const archivableOn = descriptor.archivingSchedule
-    ? formatItalianDate(descriptor.archivingSchedule.archivableOn)
+    ? dateAtRomeZone(descriptor.archivingSchedule.archivableOn)
     : undefined;
   const subject = `Archiviazione anticipata della versione ${descriptor.version} dell'e-service "${eservice.name}"`;
