+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceDescriptorArchivingScheduledToConsumer.ts
@@ -18,7 +18,7 @@ import {
 } from "pagopa-interop-notification-commons";
 import { EServiceDescriptorHandlerParams } from "../handlerCommons.js";
 import { config } from "../../config/config.js";
-import { formatItalianDate } from "./archivingDateFormat.js";
+import { dateAtRomeZone } from "pagopa-interop-commons";
 
 const notificationType: NotificationType = "eserviceStateChangedToConsumer";
 
@@ -77,7 +77,7 @@ export async function handleEserviceDescriptorArchivingScheduledToConsumer(
   }
 
   const archivableOn = descriptor.archivingSchedule
-    ? formatItalianDate(descriptor.archivingSchedule.archivableOn)
+    ? dateAtRomeZone(descriptor.archivingSchedule.archivableOn)
     : undefined;
   const subject = `Avvio archiviazione della versione ${descriptor.version} dell'e-service "${eservice.name}"`;
