+++ packages/email-notification-dispatcher/src/handlers/eservices/handleEserviceDescriptorArchivingScheduledToProducer.ts
@@ -17,7 +17,7 @@ import {
 } from "pagopa-interop-notification-commons";
 import { EServiceDescriptorHandlerParams } from "../handlerCommons.js";
 import { config } from "../../config/config.js";
-import { formatItalianDate } from "./archivingDateFormat.js";
+import { dateAtRomeZone } from "pagopa-interop-commons";
 
 const notificationType: NotificationType = "eserviceStateChangedToProducer";
 
@@ -70,7 +70,7 @@ export async function handleEserviceDescriptorArchivingScheduledToProducer(
   }
 
   const archivableOn = descriptor.archivingSchedule
-    ? formatItalianDate(descriptor.archivingSchedule.archivableOn)
+    ? dateAtRomeZone(descriptor.archivingSchedule.archivableOn)
     : undefined;
   const subject = `Avvio archiviazione della versione ${descriptor.version} dell'e-service "${eservice.name}"`;
