+++ packages/notification-commons/src/templates/email/templates.ts
@@ -100,9 +100,6 @@ export const eventMailTemplateType = {
   purposeQuotaOverthresholdMailTemplate: "purpose-quota-overthreshold-mail",
   purposeQuotaAdjustmentResponseMailTemplate:
     "purpose-quota-adjustment-response-mail",
-  asyncEserviceWithoutKeychainMailTemplate:
-    "async-eservice-without-keychain-mail",
-  // eservice archiving real-time (event-driven)
   eserviceArchivingStartedDescriptorToProducerMailTemplate:
     "eservice-archiving-started-descriptor-to-producer-mail",
   eserviceArchivingStartedDescriptorToConsumerMailTemplate:
